refactor(FE): store.tsx 파일을 index.ts로, userInfoSlice.tsx 파일을 userInfoSlice.ts로 변경
