+++ front/src/components/KakaoLogoutButton.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@/components/ui/button";
 import { logOutKakao } from "@/util/api";
 import { LogOutIcon } from "lucide-react";
-import { persistor } from "@/store/store";
+import { persistor } from "@/store/index";
 
 export default function KakaoLogoutButton() {
   const handleLogOut = async () => {
