+++ front/src/components/UserProfile.tsx
@@ -1,6 +1,6 @@
 import { Avatar, AvatarImage, AvatarFallback } from "@/components/ui/avatar";
 import { useSelector } from "react-redux";
-import { RootState } from "@/store/store";
+import { RootState } from "@/store/index";
 import { UserRound } from "lucide-react";
 
 export default function UserProfile() {
